style: Factor out some of style_derive.
Differential Revision: https://phabricator.services.mozilla.com/D17188
