+++ components/style_derive/lib.rs
@@ -6,6 +6,7 @@
 
 #[macro_use]
 extern crate darling;
+extern crate derive_common;
 extern crate proc_macro;
 extern crate proc_macro2;
 #[macro_use]
@@ -17,7 +18,6 @@ extern crate synstructure;
 use proc_macro::TokenStream;
 
 mod animate;
-mod cg;
 mod compute_squared_distance;
 mod parse;
 mod specified_value_info;
