+++ components/style_derive/parse.rs
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use crate::cg;
 use crate::to_css::CssVariantAttrs;
+use derive_common::cg;
 use proc_macro2::TokenStream;
 use syn::{self, DeriveInput, Path};
 use synstructure::{Structure, VariantInfo};
