+++ components/style_derive/to_animated_value.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use crate::cg;
+use derive_common::cg;
 use proc_macro2::{Span, TokenStream};
 use syn::{DeriveInput, Ident};
 use synstructure::BindStyle;
